Rename navbar, pull report to insights (#296)
* Allow buildly invite to register

* new version and env variable recognize changes

* Nav naming changes

* rename folders and files

* Move report to Insights

* update logos

* merge report page changes

* patch version
