+++ initialize_container.sh
@@ -4,7 +4,7 @@
 export GIT_FETCH_HEAD=`cat .git/FETCH_HEAD`
 
 #Read all env variables as output by printenv and put them into an object stored in window.env
-RESULT='window.env = {'
+RESULT='{'
 RESULT+='API_URL: "'$API_URL
 RESULT+='", OAUTH_CLIENT_ID: "'$OAUTH_CLIENT_ID
 RESULT+='", OAUTH_TOKEN_URL: "'$OAUTH_TOKEN_URL
@@ -22,6 +22,6 @@ RESULT+='", PRODUCTION: '$PRODUCTION
 RESULT+='}'
 
 PATH=`ls dist/`
-OUTPUTPATH="dist/environment.js"
+OUTPUTPATH="dist/environment.json"
 
 echo $RESULT > $OUTPUTPATH
