+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "buildly-react-template",
-  "version": "v0.7.1",
+  "version": "v0.8.0",
   "description": "Frontend Template from Buildly built using the React framework",
   "main": "src/index.js",
   "private": true,
@@ -76,7 +76,10 @@
     "typescript": "^4.4.2",
     "uuid": "3.3.3",
     "web-vitals": "^2.1.0",
-    "xstate": "^4.37.2"
+    "xstate": "^4.37.2",
+    "workbox-core": "^6.5.1",
+    "workbox-precaching": "^6.5.1",
+    "workbox-window": "^6.5.1"
   },
   "devDependencies": {
     "@babel/cli": "^7.2.3",
@@ -132,7 +135,8 @@
     "typescript": "^5.2.2",
     "webpack": "^5.85.0",
     "webpack-cli": "^3.2.1",
-    "webpack-dev-server": "^3.1.14"
+    "webpack-dev-server": "^3.1.14",
+    "workbox-webpack-plugin": "^6.5.1"
   },
   "resolutions": {
     "@babel/core": "7.19.6",
