+++ src/components/Chatbot/Chatbot.js
@@ -11,7 +11,7 @@ import {
 import SmartToyIcon from '@mui/icons-material/SmartToy';
 
 // Set sender here
-const sender = 'Insights';
+const sender = 'Buildly Product Labs';
 // "Explain things like you would to a 10 year old learning how to code."
 const systemMessage = {
   //  Explain things like you're talking to a software professional with 5 years of experience.
