--- src/components/Navbar/Navbar.css
@@ -1,27 +1,0 @@
-.navbar-brand {
-  background: url("/src/assets/insights-orange-white.png") center / contain no-repeat;
-  width: 10rem;
-  height: 2.5rem;
-}
-
-.main-nav-link .nav-link {
-  color: #ffffff !important;
-  text-transform: uppercase;
-  font-size: 15px;
-  line-height: 1.25rem;
-  align-items: center;
-  text-align: center;
-}
-
-.user-info {
-  font-size: 12px;
-  color: white;
-  display: flex;
-  flex-direction: column;
-  place-content: flex-end center;
-  align-items: flex-end;
-}
-
-.dropdown-toggle::after {
-  display: none !important;
-}
