+++ src/components/Timeline/TimelineComponent.css
@@ -1,5 +1,6 @@
 .feature-list {
-  max-height: 180px;
+  width: 210px;
+  height: 180px;
   overflow-y: auto;
   padding: 2px;
   border-radius: 4px;
