+++ src/modules/release/details/ReleaseDetails.tsx
@@ -415,7 +415,7 @@ function ReleaseDetails() {
           </Box>
         </>
       )}
-      <Chatbot />
+      {/* <Chatbot /> */}
     </>
   );
 }
