+++ src/pages/Login/Login.js
@@ -12,7 +12,7 @@ import {
   Typography,
   Container,
 } from '@mui/material';
-import logo from '@assets/insights-logo.png';
+import logo from '@assets/buildly-product-labs-logo.png';
 import Copyright from '@components/Copyright/Copyright';
 import GithubLogin from '@components/SocialLogin/GithubLogin';
 import Loader from '@components/Loader/Loader';
