+++ src/pages/NewProduct/NewProduct.js
@@ -192,7 +192,7 @@ const NewProductForm = (props) => {
   const steps = getSteps();
   const maxSteps = steps.length;
 
-  const redirectTo = (location.state && location.state.from) || routes.ROADMAP;
+  const redirectTo = (location.state && location.state.from) || routes.PRODUCT_ROADMAP;
   const product_uuid = location.state && location.state.product_uuid;
   const editPage = location.state && location.state.type === 'editP';
   const editData = (
@@ -334,7 +334,7 @@ const NewProductForm = (props) => {
               )}
             </div>
           </div>
-          <Chatbot />
+          {/* <Chatbot /> */}
         </FormModal>
       )}
     </div>
