+++ src/pages/ProductRoadmap/forms/StatusBoard.js
@@ -16,7 +16,7 @@ import {
 import FormModal from '@components/Modal/FormModal';
 import Loader from '@components/Loader/Loader';
 import { createStatus } from '@redux/release/actions/release.actions';
-import { STATUSTYPES } from '../RoadmapConstants';
+import { STATUSTYPES } from '../ProductRoadmapConstants';
 
 const useStyles = makeStyles((theme) => ({
   form: {
