+++ src/pages/ProductRoadmap/forms/ToolBoard.js
@@ -16,7 +16,7 @@ import {
 import FormModal from '@components/Modal/FormModal';
 import Loader from '@components/Loader/Loader';
 import { createBoard } from '@redux/product/actions/product.actions';
-import { STATUSTYPES } from '../RoadmapConstants';
+import { STATUSTYPES } from '../ProductRoadmapConstants';
 
 const useStyles = makeStyles((theme) => ({
   form: {
