+++ src/pages/ResetPassword/ResetPassword.js
@@ -12,7 +12,7 @@ import {
   Container,
   Grid,
 } from '@mui/material';
-import logo from '@assets/insights-logo.png';
+import logo from '@assets/buildly-product-labs-logo.png';
 import Copyright from '@components/Copyright/Copyright';
 import Loader from '@components/Loader/Loader';
 import { useInput } from '@hooks/useInput';
