+++ src/redux/authuser/sagas/authuser.saga.js
@@ -89,7 +89,7 @@ function* login(payload) {
     yield call(oauthService.setCurrentCoreUser, coreuser, user);
     yield [
       yield put({ type: LOGIN_SUCCESS, user }),
-      yield call(history.push, routes.ROADMAP),
+      yield call(history.push, routes.PRODUCT_PORTFOLIO),
     ];
   } catch (error) {
     console.log('error', error);
@@ -441,7 +441,7 @@ function* socialLogin(payload) {
     if (!user.data.email || !user.data.organization || !user.data.user_type) {
       yield call(history.push, routes.MISSING_DATA);
     } else {
-      yield call(history.push, routes.ROADMAP);
+      yield call(history.push, routes.PRODUCT_PORTFOLIO);
     }
   } catch (error) {
     yield [
@@ -523,7 +523,7 @@ function* addOrgSocialUser(payload) {
       ];
     } else if (!existingOrg && data.organization_name !== 'default organization') {
       yield [
-        yield call(history.push, routes.ROADMAP),
+        yield call(history.push, routes.PRODUCT_PORTFOLIO),
         yield put(
           showAlert({
             type: 'success',
