+++ src/redux/product/sagas/product.saga.js
@@ -413,7 +413,7 @@ function* createProduct(payload) {
     if (history) {
       // save active product uuid in local storage
       localStorage.setItem('activeProduct', product.data.product_uuid);
-      history.push(routes.ROADMAP, { selected_product: product.data.product_uuid });
+      history.push(routes.PRODUCT_ROADMAP, { selected_product: product.data.product_uuid });
     }
   } catch (error) {
     yield [
