+++ src/routes/routesConstants.js
@@ -6,10 +6,10 @@ export const routes = {
   RESET_PASSWORD: '/reset-password-confirm',
   VERIFY_EMAIL: '/verify-email',
   APP: '/app',
-  ROADMAP: '/app/roadmap',
-  ROADMAP_TABULAR: '/app/roadmap/tabular',
-  ROADMAP_KANBAN: '/app/roadmap/kanban',
-  ROADMAP_REPORT: '/app/roadmap/report',
+  PRODUCT_ROADMAP: '/app/product-roadmap',
+  PRODUCT_ROADMAP_TABULAR: '/app/product-roadmap/tabular',
+  PRODUCT_ROADMAP_KANBAN: '/app/product-roadmap/kanban',
+  PRODUCT_ROADMAP_REPORT: '/app/product-roadmap/report',
   USER_PROFILE: '/app/profile/user',
   EDIT_USER_PROFILE: '/app/profile/user/edit-profile',
   USER_SUBSCRIPTIONS: '/app/profile/user/subscriptions',
@@ -21,17 +21,18 @@ export const routes = {
   MISSING_DATA: '/missing-data',
   NEW_PRODUCT: '/app/new-product',
   RELEASE: '/app/release',
+  INSIGHTS: '/app/insights',
   DEVELOPER_FORM: '/app/developer-form',
-  PRODUCTS: '/app/products',
-  ADD_FEATURE: '/app/roadmap/add-feature/',
-  EDIT_FEATURE: '/app/roadmap/edit-feature/',
-  VIEW_FEATURE: '/app/roadmap/view-feature/',
-  ADD_ISSUE: '/app/roadmap/add-issue/',
-  EDIT_ISSUE: '/app/roadmap/edit-issue/',
-  ISSUE_SUGGESTIONS: '/app/roadmap/issue-suggestions',
-  FEATURE_TO_ISSUE: '/app/roadmap/convert-issue',
-  STATUS_BOARD: '/app/roadmap/tool-status',
-  TOOL_BOARD: '/app/roadmap/tool-board',
-  COMMENTS: '/app/roadmap/comments',
-  SHOW_RELATED_ISSUES: '/app/roadmap/show-related-issues',
+  PRODUCT_PORTFOLIO: '/app/product-portfolio',
+  ADD_FEATURE: '/app/product-roadmap/add-feature/',
+  EDIT_FEATURE: '/app/product-roadmap/edit-feature/',
+  VIEW_FEATURE: '/app/product-roadmap/view-feature/',
+  ADD_ISSUE: '/app/product-roadmap/add-issue/',
+  EDIT_ISSUE: '/app/product-roadmap/edit-issue/',
+  ISSUE_SUGGESTIONS: '/app/product-roadmap/issue-suggestions',
+  FEATURE_TO_ISSUE: '/app/product-roadmap/convert-issue',
+  STATUS_BOARD: '/app/product-roadmap/tool-status',
+  TOOL_BOARD: '/app/product-roadmap/tool-board',
+  COMMENTS: '/app/product-roadmap/comments',
+  SHOW_RELATED_ISSUES: '/app/product-roadmap/show-related-issues',
 };
