--- window.environment.js
@@ -1,1 +1,0 @@
-window.env = null;
